requirements.txt
opencv-python==4.6.0.66
numpy
Pillow